ajustes nos nomes para o padrão dart
pra tirar os avisos de erro
